+++ packages/app/src/components/swap/AddLiquidity.tsx
@@ -1,6 +1,6 @@
-import { FC, useState, ChangeEventHandler, MouseEventHandler, useEffect } from "react";
+import { FC, useState, ChangeEventHandler, MouseEventHandler } from "react";
 import { addresses, contracts } from "@tender/contracts/src/index";
-import { BigNumberish, constants, utils } from "ethers";
+import { BigNumberish, utils } from "ethers";
 import {
   Image,
   Button,
@@ -20,10 +20,10 @@ import ApproveToken from "components/approve/ApproveToken";
 import { useIsTokenApproved } from "components/approve/useIsTokenApproved";
 import { AmountInputFooter } from "components/AmountInputFooter";
 import { LoadingButtonContent } from "components/LoadingButtonContent";
-import { useCalculateLpTokenAmount, useAddLiquidity, useLiquidityPriceImpact } from "utils/tenderSwapHooks";
+import { useAddLiquidity, useLiquidityPriceImpact } from "utils/tenderSwapHooks";
 import { hasValue, useBalanceValidation } from "utils/inputValidation";
 import { isPendingTransaction } from "utils/transactions";
-import { weiToEthInFloat, weiToEthWithDecimals, withDecimals } from "utils/amountFormat";
+import { weiToEthWithDecimals, withDecimals } from "utils/amountFormat";
 import { stakers } from "@tender/shared/src/index";
 import { useEthers } from "@usedapp/core";
 import { FormClose } from "grommet-icons";
@@ -31,6 +31,7 @@ import { useResetInputAfterTx } from "utils/useResetInputAfterTx";
 import { ProtocolName } from "@tender/shared/src/data/stakers";
 import { useIsGnosisSafe } from "utils/context";
 import { SlippageInput } from "components/SlippageInput";
+import { useLPTokenOut } from "utils/useLPTokenOut";
 
 type Props = {
   protocolName: ProtocolName;
@@ -102,31 +103,11 @@ const AddLiquidity: FC<Props> = ({ protocolName, symbol, tokenBalance, tenderTok
 
   const { addLiquidity, tx: addLiquidityTx } = useAddLiquidity(protocolName, isTokenApproved, isTenderApproved);
 
-  const defaultLPTokenOut = useCalculateLpTokenAmount(
-    addresses[protocolName].tenderSwap,
-    [utils.parseEther(tenderInput || "0"), utils.parseEther(tokenInput || "0")],
-    true
+  const { lpTokenMinOut, slippage, setSlippage, suggestedSlippage } = useLPTokenOut(
+    protocolName,
+    tenderInput,
+    tokenInput
   );
-
-  const [lpTokenMinOut, setLPTokenMinOut] = useState(defaultLPTokenOut);
-
-  const sumIn = utils.parseEther(tokenInput || "0").add(utils.parseEther(tenderInput || "0"));
-  const defaultSlippage = defaultLPTokenOut.eq(constants.Zero)
-    ? 0
-    : Math.round(Math.abs((1 - weiToEthInFloat(sumIn) / weiToEthInFloat(defaultLPTokenOut)) * 100) * 100) / 100;
-  const [slippage, setSlippage] = useState(defaultSlippage);
-
-  useEffect(() => {
-    setSlippage(defaultSlippage);
-  }, [defaultSlippage]);
-
-  useEffect(() => {
-    const tokenIn = utils.parseEther(tokenInput || "0");
-    const tenderIn = utils.parseEther(tenderInput || "0");
-    const sumIn = tokenIn.add(tenderIn);
-    const outMin = sumIn.sub(sumIn.mul(slippage * 100).div(10000));
-    setLPTokenMinOut(outMin);
-  }, [slippage, tokenInput, tenderInput]);
 
   const handleAddLiquidity: MouseEventHandler<HTMLButtonElement & HTMLAnchorElement> = async (e) => {
     e.preventDefault();
@@ -226,7 +207,7 @@ const AddLiquidity: FC<Props> = ({ protocolName, symbol, tokenBalance, tenderTok
                         }
                       />
                     </FormField>
-                    <SlippageInput slippage={slippage} setSlippage={setSlippage} auto={defaultSlippage} />
+                    <SlippageInput slippage={slippage} setSlippage={setSlippage} auto={suggestedSlippage} />
                     <Text textAlign="end">{`Price impact: ${withDecimals((priceImpact * 100).toString(), 2)} %`}</Text>
                   </Box>
                 </Form>
